+++ fupashop/database/migrations/2017_09_28_035248_create_tablets_table.php
@@ -16,17 +16,17 @@ class CreateTabletsTable extends Migration
         Schema::create('tablets', function (Blueprint $table) {
             $table->string('modelNumber', 20);
             $table->string('processor', 20);
-            $table->double('screensize', 3, 1);
+            $table->decimal('screensize', 3, 1);
             $table->string('dimensions', 40);
             $table->string('ramSize', 20);
-            $table->double('weight', 5, 2);
+            $table->decimal('weight', 10, 2);
             $table->string('cpucores', 20);
             $table->string('hddSize', 20);
             $table->string('batteryInformation', 20);
             $table->string('brandName', 30);
             $table->string('operatingSystem', 20);
             $table->string('cameraInformation', 40);
-            $table->double('price', 5, 2);
+            $table->decimal('price', 10, 2);
              $table->primary('modelNumber');
         });
     }
